+++ extensions/sips/src/operations/convertOperation.ts
@@ -9,10 +9,12 @@
  */
 
 import { execSync } from "child_process";
+import { readdirSync } from "fs";
 import path from "path";
 
 import { environment, getPreferenceValues } from "@raycast/api";
 
+import { getAVIFEncPaths } from "../utilities/avif";
 import {
   addItemToRemove,
   convertPDF,
@@ -22,9 +24,34 @@ import {
   getWebPBinaryPath,
   moveImageResultsToFinalDestination,
 } from "../utilities/utils";
-import { getAVIFEncPaths } from "../utilities/avif";
-import { readdirSync } from "fs";
-import { ConvertPreferences, ExtensionPreferences } from "../utilities/preferences";
+
+/**
+ * All supported image formats for conversion.
+ */
+export const imageFormats = [
+  "ASTC",
+  "AVIF",
+  "BMP",
+  "DDS",
+  "EXR",
+  "GIF",
+  "HEIC",
+  "HEICS",
+  "ICNS",
+  "ICO",
+  "JPEG",
+  "JP2",
+  "KTX",
+  "PBM",
+  "PDF",
+  "PNG",
+  "PSD",
+  "PVR",
+  "TGA",
+  "TIFF",
+  "WEBP",
+  "SVG",
+] as const;
 
 /**
  * Converts images to the specified format, storing the results according to the user's preferences.
@@ -39,7 +66,11 @@ export default async function convert(
   outputPaths?: string[],
   intermediate = false,
 ) {
-  const preferences = getPreferenceValues<ConvertPreferences & ExtensionPreferences>();
+  const preferences = getPreferenceValues<Preferences.Convert>();
+  if (environment.commandName === "tools/convert-images") {
+    preferences.jpegExtension = "jpg";
+  }
+
   const resultPaths = [];
   for (const [index, item] of sourcePaths.entries()) {
     const originalType = path.extname(item).slice(1);
@@ -81,20 +112,17 @@ export default async function convert(
         // SVG -> PNG -> AVIF, PDF, or WebP
         await using pngFile = await getScopedTempFile("tmp", "png");
         await convertSVG("PNG", item, pngFile.path);
-        await convert([pngFile.path], desiredType, [newPath]);
-        return;
+        return await convert([pngFile.path], desiredType, [newPath]);
       } else {
         // SVG -> NSBitmapImageRep -> Desired Format
         await convertSVG(desiredType, item, newPath);
-        await convert([newPath], desiredType, [newPath]);
-        return;
+        return await convert([newPath], desiredType, [newPath]);
       }
     } else if (desiredType == "SVG") {
       await using bmpFile = await getScopedTempFile("tmp", "bmp");
       execSync(`chmod +x ${environment.assetsPath}/potrace/potrace`);
       if (originalType.toLowerCase() == "webp") {
         // WebP -> PNG -> BMP -> SVG
-        // const pngPath = await scopedTempFile("tmp", "png");
         await using pngFile = await getScopedTempFile("tmp", "png");
         const [dwebpPath] = await getWebPBinaryPath();
         execSync(`${dwebpPath} "${item}" -o "${pngFile.path}"`);
@@ -170,8 +198,7 @@ export default async function convert(
       const { decoderPath } = await getAVIFEncPaths();
       await using pngFile = await getScopedTempFile("tmp", "png");
       execSync(`${decoderPath} "${item}" "${pngFile.path}"`);
-      await convert([pngFile.path], desiredType, [newPath]);
-      return;
+      return await convert([pngFile.path], desiredType, [newPath]);
     } else {
       // General Input Format -> Desired Format
       execSync(`sips --setProperty format ${desiredType.toLowerCase()} "${item}" --out "${newPath}"`);
@@ -183,4 +210,5 @@ export default async function convert(
   if (!intermediate) {
     await moveImageResultsToFinalDestination(resultPaths);
   }
+  return resultPaths;
 }