+++ extensions/clickup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ClickUp Changelog
 
+## [✨ AI Enhancements] - 2025-02-21
+
+- AI Tools to:
+    1. Get Teams
+    2. Get Spaces
+    3. Get Docs
+
 ## [Update - New Command] - 2024-09-10
 
 - New `Docs Explorer` command - for now it shows docs and their pages