+++ extensions/coffee/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Coffee Changelog
 
+## [✨ AI Enhancements] - 2025-02-21
+
 ## [Enhancement] - 2025-01-31
 
 - Added `Caffeinate Until` command.