+++ extensions/messages/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Messages Changelog
 
+## [✨ AI Enhancements] - 2025-02-21
+
 ## [Fix `Paste Latest OTP Code` Command for Phone numbers] - 2025-02-10
 
 Added a number filter for OTP Code messages that include phone/cellular numbers that should not be used as OTP Codes for this command.