+++ extensions/safari/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Safari Changelog
 
+## [Update] - 2025-02-21
+
+AI Tools to;
+- Search Bookmarks
+- Search History
+- Search Reading List
+- Add to Reading List
+- Open (in Safari)
+
 ## [Update] - 2025-02-14
 
 - Added a preference to enable pinyin search for Chinese characters.